requirements.txt
google-genai
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
